% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_cmi.R
\name{bootstrap_cmi}
\alias{bootstrap_cmi}
\title{Bootstrap conditional mean imputation (CMI) for a censored predictor}
\usage{
bootstrap_cmi(
  imputation_model,
  analysis_model,
  transform_imp = NULL,
  W = NULL,
  Delta = NULL,
  Z = NULL,
  data,
  est_surv,
  trapezoidal_rule = FALSE,
  Xmax = Inf,
  dist = "weibull",
  surv_between = "carry-forward",
  surv_beyond = "exponential",
  useSURV,
  B = 1000
)
}
\arguments{
\item{imputation_model}{A formula input for the covariate imputation model.}

\item{analysis_model}{A formula input for the final analysis model of interest.}

\item{transform_imp}{(Optional) A function to transform the imputed values before fitting \code{analysis_model}.}

\item{W}{(Optional) Column name of observed predictor values (including censored opens).}

\item{Delta}{(Optional) Column name of censoring indicators. Note that \code{Delta = 0} is interpreted as a censored observation.}

\item{Z}{(Optional) Column name of additional fully observed covariates.}

\item{data}{Dataframe or named matrix containing columns \code{W}, \code{Delta}, and \code{Z}.}

\item{est_surv}{A string for which CMI approach to be used: fully-parametric (\code{"FP"}), Kaplan-Meier (\code{"KM"}), semiparametric (\code{"SP"}), or nonparametric (\code{"NP"}).}

\item{trapezoidal_rule}{A logical input for whether the trapezoidal rule should be used to approximate the integral in the imputed values. Default is \code{FALSE}.}

\item{Xmax}{(Optional) Upper limit of the domain of the censored predictor. Default is \code{Xmax = Inf}.}

\item{dist}{(If \code{est_surv = "FP"}) The assumed distribution for \code{W} in the AFT model, passed to \code{survival::survreg()}. Default is \code{"weibull"}.}

\item{surv_between}{(If \code{est_surv = "KM"}, \code{"SP"}, or \code{"NP"}) A string for the method to be used to interpolate for censored values between events. Options include \code{"carry-forward"} (default), \code{"linear"}, or \code{"mean"}.}

\item{surv_beyond}{(If \code{est_surv = "KM"}, \code{"SP"}, or \code{"NP"}) A string for the method to be used to extrapolate the survival curve beyond the last observed event. Options include \code{"drop-off"}, \code{"exponential"} (default), or \code{"weibull"}.}

\item{useSURV}{(If \code{est_surv = "custom"}) Assumed survival function for \code{W} given \code{Z}. The only arguments to \code{useSURV} should be \code{W} and \code{Z}, in that order.}

\item{B}{Number of bootstrap resampling replicates. Default is \code{B = 100}.}
}
\value{
A list containing the following for the coefficients of \code{analysis_model}:
\item{table}{A dataframe containing the bootstrapped standard errors and 95\% quantile intervals.}
\item{vcov}{An estimate of the covariance matrix.}
}
\description{
Bootstrap standard errors / confidence intervals for conditional mean imputation (CMI) for a censored predictor.
}
