% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmi_fp.R
\name{cmi_fp}
\alias{cmi_fp}
\title{Fully parametric conditional mean imputation (CMI) for a censored predictor}
\usage{
cmi_fp(
  W,
  Delta,
  Z,
  data,
  fit = NULL,
  dist = "weibull",
  infinite_integral = TRUE,
  maxiter = 100
)
}
\arguments{
\item{W}{character, column name for observed values of the censored covariate}

\item{Delta}{character, column name for censoring indicators. Note that \code{Delta = 0} is interpreted as a censored observation.}

\item{Z}{character vector, column name(s) of additional fully observed covariate(s).}

\item{data}{Dataframe or named matrix containing columns \code{W}, \code{Delta}, and \code{Z}.}

\item{fit}{\code{survreg} fit, imputation model object modeling \code{W} on \code{Z}. If \code{fit = NULL} (default), the AFT model with only main effects for \code{Z} and assuming distribution \code{dist} is fit internally and used.}

\item{dist}{(optional) character, assumed distribution for \code{W} in the AFT model, passed to \code{survival::survreg()}. If \code{fit} is supplied, no need to supply \code{dist}; if \code{fit} is not supplied, default is \code{dist = "weibull"}.}

\item{infinite_integral}{(optional) logical, if \code{infinite_integral = TRUE} (default) then conditional means are found by integrating from \code{W} to \code{Inf}}

\item{max_iter}{(optional) numeric, maximum iterations allowed in call to \code{survival::survreg()}. Default is \code{100}.}
}
\value{
\item{imputed_data}{A copy of \code{data} with added column \code{imp} containing the imputed values.}
\item{code}{Indicator of algorithm status (\code{TRUE} or \code{FALSE}).}
}
\description{
Fully parametric conditional mean imputation (CMI) for a censored predictor using an accelerated failure-time (AFT) model to estimate conditional survival.
}
