% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmi_fp_weibull_bootstrap.R
\name{cmi_fp_weibull_bootstrap}
\alias{cmi_fp_weibull_bootstrap}
\title{Multiple, fully parametric conditional mean imputation for a censored covariate (Weibull distribution)}
\usage{
cmi_fp_weibull_bootstrap(
  imputation_formula,
  analysis_formula,
  W,
  Delta,
  Z,
  data,
  infinite_integral = TRUE,
  maxiter = 100,
  B = 10
)
}
\arguments{
\item{imputation_formula}{imputation model formula (or coercible to formula), a formula expression as for other regression models. The response is usually a survival object as returned by the \code{Surv} function. See the documentation for \code{Surv} for details.}

\item{analysis_formula}{analysis model formula (or coercible to formula), a formula expression as for other regression models. The response should be a continuous random variable in \code{data}.}

\item{W}{character, column name for observed values of the censored covariate}

\item{Delta}{character, column name for censoring indicators. Note that \code{Delta = 0} is interpreted as a censored observation.}

\item{Z}{character vector, column name(s) of additional fully observed covariate(s).}

\item{data}{Dataframe or named matrix containing columns \code{W}, \code{Delta}, and \code{Z}.}

\item{infinite_integral}{(optional) logical, if \code{infinite_integral = TRUE} (default) then conditional means are found by integrating from \code{W} to \code{Inf}, whereas if \code{infinite_integral = FALSE} they are found by subtracting the integral from \code{0} to \code{W} from the mean. If \code{infinite_integral = NA} instead, the analytical solutions are used to find the conditional means.}

\item{B}{numeric, number of imputations. Default is \code{10}.}

\item{max_iter}{(optional) numeric, maximum iterations allowed in call to \code{survival::survreg()}. Default is \code{100}.}
}
\value{
\item{imputed_data}{A copy of \code{data} with added column \code{imp} containing the imputed values.}
\item{code}{Indicator of algorithm status (\code{TRUE} or \code{FALSE}).}
}
\description{
Multiple, fully parametric conditional mean imputation for a censored covariate using an accelerated failure-time model with a Weibull distribution to estimate the conditional survival function.
}
