% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmi_np.R
\name{cmi_np}
\alias{cmi_np}
\title{Nonparametric conditional mean imputation (CMI) for a censored predictor}
\usage{
cmi_np(
  W,
  Delta,
  Z,
  data,
  surv_between = "carry-forward",
  surv_beyond = "exponential"
)
}
\arguments{
\item{W}{Column name of observed predictor values (including censored opens).}

\item{Delta}{Column name of censoring indicators. Note that \code{Delta = 0} is interpreted as a censored observation.}

\item{Z}{Column name of additional fully observed covariates.}

\item{data}{Dataframe or named matrix containing columns \code{W}, \code{Delta}, and \code{Z}.}

\item{surv_between}{A string for the method to be used to interpolate for censored values between events. Options include \code{"carry-forward"} (default), \code{"linear"}, or \code{"mean"}.}

\item{surv_beyond}{A string for the method to be used to extrapolate the survival curve beyond the last observed event. Options include \code{"drop-off"}, \code{"exponential"} (default), or \code{"weibull"}.}
}
\value{
A copy of \code{data} with added column \code{imp} containing the imputed values.
}
\description{
Nonparametric conditional mean imputation (CMI) for a censored predictor using a modified Nelson-Aalen estimator (STRIDE) to estimate conditional survival.
}
