% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmi_sp_resample.R
\name{cmi_sp_resample}
\alias{cmi_sp_resample}
\title{Semiparametric conditional mean imputation (CMI) for a censored predictor}
\usage{
cmi_sp_resample(
  imputation_model,
  data,
  M = 1,
  trapezoidal_rule = FALSE,
  Xmax = Inf,
  surv_between = "cf",
  surv_beyond = "e"
)
}
\arguments{
\item{imputation_model}{Imputation model formula (or coercible to formula), a formula expression as for other regression models. The response is usually a survival object as returned by the \code{Surv} function. See the documentation for \code{Surv} for details.#' @param data Dataframe or named matrix containing columns \code{W}, \code{Delta}, and \code{Z}.}

\item{M}{Number of times to resample Cox model coefficients}

\item{trapezoidal_rule}{A logical input for whether the trapezoidal rule should be used to approximate the integral in the imputed values. Default is \code{FALSE}.}

\item{Xmax}{(Optional) Upper limit of the domain of the censored predictor. Default is \code{Xmax = Inf}.}

\item{surv_between}{A string for the method to be used to interpolate for censored values between events. Options include \code{"cf"} (carry forward, the default), \code{"wm"} (weighted mean), or \code{"m"} (mean).}

\item{surv_beyond}{A string for the method to be used to extrapolate the survival curve beyond the last observed event. Options include \code{"d"} (immediate drop off), \code{"e"} (exponential extension, the default), or \code{"w"} (weibull extension).}
}
\value{
\item{imputed_datasets}{An list of length m, containing copies of \code{data} each with added column \code{imp} containing the imputed values.}
}
\description{
Semiparametric conditional mean imputation (CMI) for a censored predictor using a Cox proportional hazards model and the Breslow estimator to estimate conditional survival. 
Regression coefficients from the Cox model are resampled from their estimated multivariate normal distribution.
}
