% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csi_w_boot.R
\name{csi_w_boot}
\alias{csi_w_boot}
\title{Semiparametric single conditional mean imputation (CMI) with boostrapping for a censored predictor with}
\usage{
csi_w_boot(
  imputation_model,
  analysis_model,
  data,
  integral = "AQ",
  Xmax = Inf,
  subdivisions = 100L,
  surv_between = "cf",
  surv_beyond = "e",
  B = 500,
  stratify = FALSE
)
}
\arguments{
\item{imputation_model}{Imputation model formula (or coercible to formula), a formula expression as for other regression models. The response is usually a survival object as returned by the \code{Surv} function. See the documentation for \code{Surv} for details.}

\item{analysis_model}{Analysis model formula (or coercible to formula), a formula expression as for other regression models. The response should be a continuous outcome for normal linear regression.}

\item{data}{Dataframe or named matrix containing columns \code{W}, \code{Delta}, and \code{Z}.}

\item{integral}{A string input for how to approximate the integral in the imputed values. Default is \code{integral="AQ"} for adaptive quadrature, but \code{"TR"} (trapezoidal rule) and \code{"A"} (quasi-analytical) are also available.}

\item{Xmax}{(Optional) Upper limit of the domain of the censored predictor. Default is \code{Xmax = Inf}.}

\item{subdivisions}{(Optional) Passed through to \code{integrate}, the maximum number of subintervals. Default is \code{subdivisions = 100L}.}

\item{surv_between}{A string for the method to be used to interpolate for censored values between events. Options include \code{"cf"} (carry forward, the default), \code{"wm"} (weighted mean), or \code{"m"} (mean).}

\item{surv_beyond}{A string for the method to be used to extrapolate the survival curve beyond the last observed event. Options include \code{"d"} (immediate drop off), \code{"e"} (exponential extension, the default), or \code{"w"} (weibull extension).}

\item{B}{Numeric, number of bootstraps used for standard errors Default is \code{500}.}

\item{stratify}{Logical, if \code{stratify = FALSE} (the default), the bootstraps resample from the pooled sample. If \code{stratify = TRUE}, the bootstraps resample separately from the censored and uncensored subsamples.}
}
\value{
a \code{dataframe} of pooled coefficient and standard error estimates
}
\description{
Semiparametric single conditional mean imputation (CMI) with bootstrapping for a censored predictor using a Cox proportional hazards model and the Breslow estimator to estimate conditional survival.
}
