% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weib_params.R
\name{get_weib_params}
\alias{get_weib_params}
\title{Constrained maximum likelihood estimators for the Weibull extensions}
\usage{
get_weib_params(imputation_model, data, Xmax = Inf)
}
\arguments{
\item{imputation_model}{Imputation model formula (or coercible to formula), a formula expression as for other regression models. The response is usually a survival object as returned by the \code{Surv} function. See the documentation for \code{Surv} for details.}

\item{data}{Dataframe or named matrix containing columns \code{W}, \code{Delta}, and \code{Z}.}

\item{Xmax}{(Optional) Upper limit of the domain of the censored predictor. Default is \code{Xmax = Inf}.}
}
\value{
A vector of parameters
}
\description{
Constrained maximum likelihood estimators for the Weibull extensions
}
