% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmi_sp.R
\name{cmi_sp}
\alias{cmi_sp}
\title{Semiparametric conditional mean imputation (CMI) for a censored predictor}
\usage{
cmi_sp(
  imputation_formula,
  W,
  Delta,
  Z,
  data,
  trapezoidal_rule = FALSE,
  Xmax = Inf,
  surv_between = "cf",
  surv_beyond = "e"
)
}
\arguments{
\item{imputation_formula}{Formula specification for the Cox proportional hazards imputation model to be fit.}

\item{W}{Column name of observed (possibly censored) predictor values.}

\item{Delta}{Column name of censoring indicators. Note that \code{data[, Delta] = 0} is interpreted as a censored observation.}

\item{Z}{Column name of additional fully observed covariates.}

\item{data}{Dataframe or named matrix containing columns \code{W}, \code{Delta}, and \code{Z}.}

\item{trapezoidal_rule}{A logical input for whether the trapezoidal rule should be used to approximate the integral in the imputed values. Default is \code{FALSE}.}

\item{Xmax}{(Optional) Upper limit of the domain of the censored predictor. Default is \code{Xmax = Inf}.}

\item{surv_between}{A string for the method to be used to interpolate for censored values between events. Options include \code{"cf"} (carry forward, the default), \code{"wm"} (weighted mean), or \code{"m"} (mean).}

\item{surv_beyond}{A string for the method to be used to extrapolate the survival curve beyond the last observed event. Options include \code{"d"} (immediate drop off), \code{"e"} (exponential extension, the default), or \code{"w"} (weibull extension).}
}
\value{
\item{imputed_data}{A copy of \code{data} with added column \code{imp} containing the imputed values.}
\item{code}{Indicator of algorithm status (\code{TRUE} or \code{FALSE}).}
}
\description{
Semiparametric conditional mean imputation (CMI) for a censored predictor using a Cox proportional hazards model and the Breslow estimator to estimate conditional survival.
}
