% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmi_sp.R
\name{cmi_sp}
\alias{cmi_sp}
\title{Semiparametric conditional mean imputation (CMI) for a censored predictor}
\usage{
cmi_sp(
  W,
  Delta,
  Z,
  data,
  fit = NULL,
  stratified = FALSE,
  trapezoidal_rule = FALSE,
  surv_between = "carry-forward",
  surv_beyond = "exponential"
)
}
\arguments{
\item{W}{Column name of observed predictor values (including censored opens).}

\item{Delta}{Column name of censoring indicators. Note that \code{data[, Delta] = 0} is interpreted as a censored observation.}

\item{Z}{Column name of additional fully observed covariates.}

\item{data}{Dataframe or named matrix containing columns \code{W}, \code{Delta}, and \code{Z}.}

\item{fit}{(Optional) A \code{coxph} imputation model object modeling \code{W} on \code{Z}. If \code{fit = NULL} (default), the Cox model with only main effects for \code{Z} is used.}

\item{stratified}{If \code{TRUE}, stratification in \code{W} is used to construct time-varying coefficients. Default is \code{FALSE}.}

\item{trapezoidal_rule}{A logical input for whether the trapezoidal rule should be used to approximate the integral in the imputed values. Default is \code{FALSE}.}

\item{surv_between}{A string for the method to be used to interpolate for censored values between events. Options include \code{"carry-forward"} (default), \code{"linear"}, or \code{"mean"}.}

\item{surv_beyond}{A string for the method to be used to extrapolate the survival curve beyond the last observed event. Options include \code{"drop-off"}, \code{"exponential"} (default), or \code{"weibull"}.}
}
\value{
\item{imputed_data}{A copy of \code{data} with added column \code{imp} containing the imputed values.}
\item{code}{Indicator of algorithm status (\code{TRUE} or \code{FALSE}).}
\item{splits}{(If \code{stratified = TRUE}) The number of strata used.}
}
\description{
Semiparametric conditional mean imputation (CMI) for a censored predictor using a Cox proportional hazards model and the Breslow estimator to estimate conditional survival.
}
